lisp: address driver-substrate review (shift UB, DMA leak/zeroing)
Code-reviewer fixes on the driver substrate:
- arithmetic-shift: shift in unsigned (two's-complement result) to avoid
  signed-left-shift UB on negative values / shifts into the sign bit, which
  UBSAN would otherwise trap.
- dma-alloc: physmem_alloc ignores its flags, so the pages are NOT zeroed --
  memset the mapping before returning (a DMA descriptor ring needs clean
  memory); and physmem_free the pages on the (OOM) error path instead of
  leaking physical memory.
- Removed the dead `virt == 0` checks after vmem_phystovirt (it PANICs on a bad
  address rather than returning 0) and documented that an out-of-range map is
  fatal until the capability model gates these prims.
- region-ref/region-set! (the DSL) now reject a size that is not 1/2/4/8 with a
  clear error instead of silently issuing a u64 access.
- Noted that bytes-phys truncates a physical address >= 2^61 (fixnum range);
  fine for current targets.

Host suite green; in-OS DMA round-trip self-test still passes; both cores
online.

+++ libs/lisp/src/prims.c
@@ -1062,8 +1062,8 @@ static lisp_value prim_ashift(lisp_value *a, int n, const char **e) {
         return prim_err(e, "arithmetic-shift: expects (value count)");
     int64_t v = lisp_fixnum_val(a[0]);
     int64_t s = lisp_fixnum_val(a[1]);
-    if (s >= 0)
-        return lisp_fixnum(s >= 64 ? 0 : (v << s));
+    if (s >= 0)  // shift in unsigned to avoid signed-left-shift UB; two's-complement result
+        return lisp_fixnum(s >= 64 ? 0 : (int64_t)((uint64_t)v << s));
     int64_t sh = -s;
     return lisp_fixnum(sh >= 63 ? (v < 0 ? -1 : 0) : (v >> sh));
 }
@@ -1121,6 +1121,8 @@ static lisp_value prim_bytes_length(lisp_value *a, int n, const char **e) {
 }
 
 // (bytes-phys b) -> the physical address of a DMA region (0 for a plain buffer).
+// NB: the result is a fixnum (62 signed bits), so a physical address >= 2^61 is
+// truncated -- fine for current targets (phys addrs are well under that).
 static lisp_value prim_bytes_phys(lisp_value *a, int n, const char **e) {
     if (n != 1 || !lisp_is_bytes(a[0]))
         return prim_err(e, "bytes-phys expects a byte buffer");
